Fix Git review back navigation from mobile sheet
- Replace the Git sheet with review so Back returns to the thread
- Inline app variant config and remove unused Play Store assets
