+++ apps/mobile/app.config.ts
@@ -1,7 +1,8 @@
 import type { ExpoConfig } from "expo/config";
 
 import { loadRepoEnv } from "../../scripts/lib/public-config.ts";
-import { APP_VARIANT_CONFIG, resolveAppVariant } from "./app-variants.ts";
+
+type AppVariant = "development" | "preview" | "production";
 
 const repoEnv = loadRepoEnv();
 Object.assign(process.env, repoEnv);
@@ -22,7 +23,67 @@ if (
   );
 }
 
-const variant = APP_VARIANT_CONFIG[APP_VARIANT];
+const DEVELOPMENT_ASSETS = {
+  appIcon: "./assets/splash-icon-dev.png",
+  iosIcon: "./assets/icon-composer-dev.icon",
+  splashIcon: "./assets/splash-icon-dev.png",
+  androidAdaptiveForeground: "./assets/android-icon-dev-foreground.png",
+  androidAdaptiveBackgroundColor: "#00639B",
+  androidMonochromeIcon: "./assets/android-icon-mark.png",
+  androidNotificationIcon: "./assets/android-notification-icon.png",
+  androidNotificationColor: "#00639B",
+} as const;
+
+const RELEASE_ASSETS = {
+  appIcon: "./assets/splash-icon-prod.png",
+  iosIcon: "./assets/icon-composer-prod.icon",
+  splashIcon: "./assets/splash-icon-prod.png",
+  androidAdaptiveForeground: "./assets/android-icon-mark.png",
+  androidAdaptiveBackgroundColor: "#000000",
+  androidMonochromeIcon: "./assets/android-icon-mark.png",
+  androidNotificationIcon: "./assets/android-notification-icon.png",
+  androidNotificationColor: "#FFFFFF",
+} as const;
+
+const VARIANT_CONFIG = {
+  development: {
+    appName: "T3 Code Dev",
+    scheme: "t3code-dev",
+    iosBundleIdentifier: "com.t3tools.t3code.dev",
+    androidPackage: "com.t3tools.t3code.dev",
+    relyingParty: "clerk.t3.codes",
+    assets: DEVELOPMENT_ASSETS,
+  },
+  preview: {
+    appName: "T3 Code Preview",
+    scheme: "t3code-preview",
+    iosBundleIdentifier: "com.t3tools.t3code.preview",
+    androidPackage: "com.t3tools.t3code.preview",
+    relyingParty: "clerk.t3.codes",
+    assets: RELEASE_ASSETS,
+  },
+  production: {
+    appName: "T3 Code",
+    scheme: "t3code",
+    iosBundleIdentifier: "com.t3tools.t3code",
+    androidPackage: "com.t3tools.t3code",
+    relyingParty: "clerk.t3.codes",
+    assets: RELEASE_ASSETS,
+  },
+} as const;
+
+function resolveAppVariant(value: string | undefined): AppVariant {
+  switch (value) {
+    case "development":
+    case "preview":
+    case "production":
+      return value;
+    default:
+      return "production";
+  }
+}
+
+const variant = VARIANT_CONFIG[APP_VARIANT];
 
 const dmSansFonts = {
   regular: "@expo-google-fonts/dm-sans/400Regular/DMSans_400Regular.ttf",
