+++ apps/mobile/src/features/threads/git/git-overview-navigation.ts
@@ -1,0 +1,5 @@
+export function resolveGitOverviewReviewNavigationAction(
+  presentation: "sheet" | "inspector",
+): "replace" | "navigate" {
+  return presentation === "sheet" ? "replace" : "navigate";
+}
